DPU/requirements.txt
yfinance>=0.2.46
psycopg2-binary>=2.9.10
python-dotenv>=1.0.1
APScheduler
pytz
